requirements.txt
youtube-transcript-api>=0.6.1
crewai>=0.11.0
python-dotenv>=1.0.0
openai>=1.3.0
langchain>=0.1.0
pytube>=15.0.0 